Tarefa1

Description

Introdução

A tarefa 1 baseia-se na geração randomizada de labirintos válidos. Para que um labirinto seja considerado válido, o mesmo tem de seguir as seguintes diretrizes: * O labirinto tem de ser rodeado por peças Wall (exceto no túnel);

A faixa mais central do labirinto tem de formar um túnel;

Todos os corredores do labirinto têm o mesmo comprimento;

A casa dos fantasmas está localizada no centro do labirinto.

Objetivos

Os objetivos desta tarefa passam pela implementação de um mecanismo de geração de labirintos, através da função "generateRandoms" que recebe o número de elementos a gerar, uma semente e produz uma lista de valores pseudo-aleatórios (entre 0 e 99), que posteriormente serão convertidos em peças do jogo. Esta função deve contemplar outras que permitam a geração de labirintos válidos.

Estratégia de Implementação

A estratégia utilizada nesta tarefa foi a realização de cada uma das directrizes dos labirintos válidos. Começou-se por adicionar as paredes que rodeam o labirinto, sendo que posteriormente, foram criados os túneis, através de cálculos que permitissem encontrar o centro da altura do labirinto. Para além disso, gerou-se as possíveis hipóteses para as casas dos fantasmas e através de funções que permitissem encontrar os locais exatos da mesma no labirinto a implementassem nesse local. Numa primeira instância, seriam colocados os corredores que não são alterados com a introdução da casa dos fantasmas e depois corredores apenas com a casa dos fantasmas. Através de outras funções recursivas colocou-se nestes últimos corredores (mais pequenos que os iniciais) as peças que faltavam de ambos os lados, extraídas do labirinto completo original.

Conclusão

Desta forma, é possível concluir que esta tarefa foi concretizada com muito sucesso, considerando que foi possível de implementar uma estratégia geral de implementação da casa dos fantasmas que permite introduzi-la no local correto em qualquer labirinto, independentemente das suas dimensões. Assim, os objetivos foram cumpridos com excelência.

Documentation

sampleMaze :: Maze

Sample of a maze

generateMaze

Arguments

:: Int
width of the maze

-> Int
height of the maze

-> Int
seed for random generation

-> Maze

generateMaze: this function generates the maze given a width, height and a seed

geneRandoms :: Int -> Int -> [Int]

geneRandoms: this function generates a list of random numbers

convertPiece

Arguments

:: Int
0 <= n <= 99

-> Piece

convertPiece: this function assigns a piece to a specific group of numbers

numbers2pieces :: [Int] -> [Piece]

numbers2pieces: this function converts random numbers into pieces

list2maze :: Int -> [Piece] -> Maze

list2maze: this function creates the corridors by spliting the whole list of pieces

twoWalls

Arguments

:: Int
width of the maze

-> [Piece]
all this pieces are walls

twoWalls: this function adds the top wall and the bottom wall

verticalWalls

Arguments

:: Int
width of the corridor

-> Maze

-> Maze

verticalWalls: this function adds the walls along the vertical sides of the maze

tunnel

Arguments

:: Int
the maze's height

-> Int
the maze's middle

-> Maze

-> Maze

tunnel: this function adds the tunnel(s)

tunnelEven :: Int -> Int -> Maze -> Maze

tunnelEven: auxiliary function to tunnel used if the height of the maze is an even number

tunnelOdd :: Int -> Int -> Maze -> Maze

tunnelOdd: auxiliary function to tunnel used if the height of the maze is an odd number

ghosthouse

Arguments

:: Int
the maze's width

-> Maze

ghosthouse: this function generates the ghost's house

ghousecor :: Maze -> Maze

ghousecor: this function takes the first x corridors that don't change with the insertion of the ghost house

ghousecor1 :: Maze -> Maze

ghousecor1: this function drops the last x corridors that don't change with the insertion of the ghost house

ghouseposition

Arguments

:: Maze

-> Int
the maze's width

-> Maze

ghouseposition: this function adds the results of the last two functions ghousecor and ghousecor1 with the one that generates the ghost house ghosthouse

firstpositions

Arguments

:: Int
width of the maze

-> Maze
maze with the ghost house

-> Maze
inicial maze without the ghost house

-> Maze

firstpositions: this function takes the first pieces of an even corridor and adds them to the places where they are missing

lastpositions

Arguments

:: Int
the maze's width

-> Maze
maze with the ghost house

-> Maze
inicial maze without the ghost house

-> Maze

lastpositions: this function takes the last pieces of the corridor and adds them to the places where they are missing

printMaze :: Maze -> IO ()

printMaze: this function is used to show the maze in it's game state in the terminal

printMaze' :: Maze -> String

printMaze' : auxiliary function to printMaze that converts a maze into a string

printCorridor :: Corridor -> String

printCorridor: auxiliary function to printMaze' that converts a corridor into a string