Tarefa2

Description

Introdução

A tarefa2 baseia-se na implementação de jogadas para os jogadores do tipo Pacman.

Objetivos

O objetivo desta tarefa é a implementação de jogadas do jogador Pacman, sendo que dada uma Play numa determinada direção, o estado do jogo deve ser atualizado e por sua vez, este deve atualizar o efeito dessa jogada no labirinto atual. No que à movimentação concerne, é necessário concretizar as jogadas nas diferentes direções, os casos em que nada acontece, a soma de pontos (aquando encontro do Pacman com a comida), e ainda o estado Mega do Pacman.

Estratégia de Implementação

Para esta tarefa, foram implementadas funções que permitissem determinar o tipo de peça que se encontra imediatamente após o Pacman, numa determinada direção. Desta forma, foi possível cobrir todos os casos referidos acima que englobassem o encontro com algum tipo de peça ou jogador. Para além desta, foram implementadas funções que mudassem as peças que se encontravam anteriormente nas coordenadas do Pacman, numa jogada imediatamente anterior, para peças Empty. Posteriormente, e de uma forma geral, foi executada uma função "move" que cobrisse todos os movimentos e encontros, para além de soma de pontos e perda de vidas, em todas as direções. Para a concretização do estado Mega, foi realizada uma função para este efeito (auxiliar às funções auxiliares da "move") bem como outras auxiliares a esta, que permitissem cobrir todas as exigências deste estado (encontro com fantasma morto, este volta para o meio do labirinto - local onde se encontra a casa dos fantasmas, entre outros)

Conclusão

É possível concluir, assim, que esta tarefa foi implementada com sucesso, considerando que todos os casos incluídos no enunciado foram contemplados na função principal "play", tendo sido os objetivos para esta tarefa concluídos com sucesso.

Documentation

state' :: State

state: example of a maze with a predetermined state

positionsa :: Coords -> Maze -> Maybe Piece

positionsa: this function returns a Piece placed in the given coordinates

positionsb :: Corridor -> Int -> Maybe Piece

positionsb: auxiliary function to positionsa that finds the right coordinate in a Corridor

changeToEmpty :: Coords -> Maze -> Maze

changeToEmpty: this funcion changes whatever piece placed in certain coordinates to the Piece Empty

changeToEmptyA :: Coords -> Maze -> Maze

changetoEmptyA: auxiliary function to changetoEmpty that finds the correspondent piece in a maze

changeToEmptyB :: Int -> Corridor -> Corridor

changeToEmptyB: auxiliary function to changeToEmptyA that finds the correspondent piece in a maze

pl :: [Player]

'Game Players': made to test the game

player1 :: Player

player2 :: Player

player3 :: Player

q :: State

x :: Maze

maze' :: State -> Maze

y :: Maze

play :: Play -> State -> State

play: main function that given a play (movement of the player Pacman) and a state returns another state with the changes incorporated

move :: Orientation -> State -> State

move: auxiliary function to play that manages the Pacman's movements

moveMR :: Coords -> Maze -> Maze

moveMR: auxiliary function to move that updates the maze when there are movements made to the right

moveR :: Maze -> [Player] -> [Player]

moveR: auxiliary function to move that moves the player to the right and contains all of his plays

moveML :: Coords -> Maze -> Maze

moveML: auxiliary function to move that updates the maze when there are movements to the left

moveL :: Maze -> [Player] -> [Player]

moveL: auxiliary funtion to move that moves the player to the left and contains all of his plays

moveMU :: Coords -> Maze -> Maze

moveMU: auxiliary function to move that updates the maze when movements upwards are made

moveU :: Maze -> [Player] -> [Player]

moveU: auxiliary function to move that moves the player upwards and contains all of his plays

moveMD :: Coords -> Maze -> Maze

moveMD: auxiliary funtion to move that updates the maze when movements downwards are made

moveD :: Maze -> [Player] -> [Player]

moveD: auxiliary funtion to move that moves the player downwards and contains all of his plays

megaPacmanState :: [Player] -> [Player]

megaPacmanState: this function regulates the Pacman's Mega mode

megaGhostState :: Player -> [Player] -> [Player]

megaGhostState: auxiliary funtion to "megaPacmanState" that regulates the Ghosts when Pacman is in the Mega mode

confrontation :: Player -> [Player] -> Bool

confrontation: this function returns True if there is a ghost in the same position as Pacman

onlyGhost :: [Player] -> [Player]

onlyGhost: this function only returns a list with all the ghosts players

onlyPacman :: [Player] -> [Player]

onlyPacman: this function only returns a list with all the Pacman players

onlyOneGhost :: Player -> [Player] -> Player

onlyOneGhost: this function returns a ghost with specific coordinates (same as Pacman)

middleMaze :: Maze -> Coords

middleMaze: this function returns the middle of the maze (inside the house of ghosts)

pacmanVSghost :: Maze -> [Player] -> [Player]

pacmanVSghost: this function deals with the confrontation between the pacman and the ghost

otherPlayers :: Player -> Player -> [Player] -> [Player]

otherPlayers: this function returns a list without the players that are in the same coordinates (Pacman and Ghost)

mouthPacman :: Mouth -> Mouth

mouthPacman : This function switches the Pacman's mouth between open and closed