Tarefa4

Description

Introdução

A tarefa 4 baseia-se na implementação da reação à passagem do tempo e para além disso, na implementação de movimentação dos jogadores numa interação. Desta forma, esta tarefa é divida em duas partes:

Parte 1: integra os tipos e as funções que advém da 1ª fase numam biblioteca denominada NCurses

Parte 2: utiliza o estado e informação vinda da primeira parte para avançar o jogo numa iteração, fazendo, desta forma, a movimentação de todos os jogadores

Objetivos

Os objetivos desta tarefa passam pelo cálculo do efeito da passagem de um instante de tempo num determinado estado de jogo. Através de um data Manager, no final desta tarefa, pretende-se ser possível controlar a passagem do tempo definido para efetuar uma nova jogada, que por si só, deve movimentar os jogadores, tendo em conta a sua velocidade e o respetivo estado. Para este efeito, devem ser introduzidos na tarefa 2 comportamentos afetos ao estado Mega, ao próprio Pacman (abrir e fechar a boca) e à velocidade dos jogadores.

Estratégia de Implementação

Numa primeira instância, foram introduzidas no ficheiro Main.hs algumas funções que permitissem a passagem do tempo a cada jogada, tal como, a "nextFrame", a "updateTime" e a "resetTimer". Para além disso, foi concretizada uma funçãom "updateControlledPlayer" que através das setas do teclado, permite mudar a direção do jogador Pacman. Na segunda parte desta tarefa, foi implementada a função "passTime". Esta função permitiu não só a implementação da função "ghostPlay" tarefa 5, como também o bot tarefa 6. A estratégia utilizada para esta função passou pela utilização de uma função auxiliar que cobrisse todos os casos obrigatórios nesta tarefa: a constante movimentação do Pacman na direção em que se encontra (até ser pressionada uma Key numa direção diferente), as diferentes velocidades dos jogadores (nomeadamente os fantasmas) e a implementação da tarefa 5, com a atribuição de cada Play da lista originada na função "ghostPlay" ao respetivo fantasma. A função em formato de comentário permite o funcionamento do bot num jogo Pacman.

Conclusão

De uma forma geral, pode-se concluir que esta tarefa se apresentou como desafiadora e em certas circunstâncias difícil de implementar. A prestação total desta tarefa poderia ser muito melhor contudo, apesar de algumas imperfeições e pontos falhados (nomeadamente no estado Mega), foi possível ter uma melhor compreensão do que é a passagem do tempo num jogo. A estratégia utilizada poderá não ter sido a melhor, contudo e de uma forma geral, foi possível colocar o jogo a funcionar de maneira razoável. Um dos pontos mais marcantes na realização desta tarefa foi que a mesma se demonstrou ser desafiadora.

Documentation

defaultDelayTime :: Integer

passTime :: Int -> State -> State

passTimeAux :: Int -> State -> [Player] -> [Play] -> State