Tarefa5

Description

Introdução

A tarefa 5 baseia-se na implementação da movimentação e comportamento dos jogadores fantasmas. É, desta forma, necessária a implementação da movimentação dos fantasmas na tarefa 2 e em cada instante, é necessário ter em consideração o estado de cada fantasma. Deverão ser implementadas duas funções "chaseMode" e "scatterMode", nas quais se recebe um estado do jogo e um identificador de fantasma de modo a devolver o um (Move identificador orientacao), necessário na implementãção de uma jogada Play.

Objetivos

Para a concretização desta tarefa, pretende-se que o comportamento normal, no qual o fantasma se encontra em modo Alive, seja perseguir o Pacman. Caso o fantasma esteja em modo Dead, este deve fugir ao Pacman em modo Mega até que voltar a estar em modo Alive. Adicionalmente, sempre que os fantasmas passem para o modo Dead, estes devem obter uma orientação oposta à que tinha no momento anterior. Para isto, é necessária a implementação da função "ghostPlay" que recebe um estado de jogo e devolve um conjunto de jogadas, uma para cada fantasma, com a melhor alternativa de reação ao Pacman.

Estratégia de Implementação

As estratégias utilizadas nesta tarefa, são as seguintes:

Fantasmas (Alive) em perseguição do Pacman: quando um fantasma se deparar com um muro, este irá determinar a posição do Pacman, sendo que o fantasma irá na direção mais próxima do Pacman. Por exemplo: o fantasma encontra um muro na posição imediatamente acima da sua - se o Pacman se encontrar abaixo, o fantasma seguirá essa direção, utilizando esta estratégia para as outras direções. Para além desta estratégia, colocou-se a hipótese do fantasma se encontrar entre duas paredes, tendo sido contemplada esta opção no código abaixo. Neste caso, o fantasma muda de direção antes de ficar encurralado.

Fantasmas (Dead) fogem do Pacman (Mega): quando o fantasma se depara com uma parede, este move-se para a sua direita, circulando a área onde se encontra no sentido dos ponteiros de relógico, tal como mencionado acima. Esta estratégia foi conseguida através da função "scatterMode" que sempre que muda a direção dos fantasmas como descrito.

Conclusão

Em geral, conclui-se que esta tarefa foi implementada com sucesso e que os objetivos foram todos cumpridos com excelência. Mais uma vez a qualidade do código poderia melhorar, mas o foco desta tarefa foi implementar duas estratégias que complicassem o jogo e, ao mesmo tempo, que permitissem ao jogador Pacman criar uma estratégia para tentar prever os movimentos dos fantasmas e, desta forma, ganhar o jogo.

Documentation

chaseMode

Arguments

:: State

-> Int
The ID of the ghost that will be moved

-> Play

chaseMode: This function gives movements to the ghosts so that they chase Pacman

scatterMode

Arguments

:: State

-> Int
The ID of the ghost that will be moved

-> Play

scatterMode: This function gives movements to ghosts so that they can escape Pacman

chaseMode'

Arguments

:: State

-> [Int]
list with the living ghosts' ids

-> [Play]

chaseMode: This function uses the function chaseMode to give a list of movements to go in the direction of Pacman, one for each Ghost

scatterMode'

Arguments

:: State

-> [Int]
list with the dead ghosts' ids

-> [Play]

scatterMode: This function uses the function scatterMode to give a list of movements to escape Pacman, one for each Ghost

separateAliveID

Arguments

:: [Player]
only ghosts

-> [Int]

separateAliveID: This function gives all the living ghosts IDs

separateDeadID

Arguments

:: [Player]
only ghosts

-> [Int]

separateDeadID: This function gives all dead ghosts' IDs

ghostPlay :: State -> [Play]

ghostPlay: This function receives the game state and returns a set of moves, one from each ghost