Tarefa6

Description

Introdução

A tarefa 6 consiste na criação de um robô, ou seja, de um jogador Pacman capaz de jogar automaticament e sem input humano.

Objetivos

Os objetivos desta tarefa passam pela implementação de um robô que consiga tomar decisões utilizando o Play definido. A função principal desta tarefa bot recebe um identificador de jogador e um estado de jogo e retorna uma Play do tipo Maybe.

Estratégia de Implementação

Nesta tarefa foram implementadas duas ações do bot, sendo elas:

Pacman em modo Mega persegue os fantasmas: nesta ação, o Pacman determina as coordenadas do fantasma mais próximo e vai em direção a ele. Se a distância horizontal entre os dois for menor do que a distância vertical, ele move-se para a esquerda ou para a direita, consoante a direção do fantasma. No caso contrário, o Pacman move-se para cima ou para baixo, consoante a direção em que o fantasma se encontre. Para além disso, se o Pacman se encontrar perto de um túnel e um dos fantasmas estiver perto do outro túnel, o primeiro irá se deslocar em direção ao segundo. Importa ainda referir que se o Pacman encontrar uma parede, ele irá se desviar do mesmo.

Pacman em modo Normal foge dos fantasmas: nesta ação, quando o Pacman se encontra em perigo, ou seja, perto de um dos fantasmas (3 peças vertical ou horizontalmente), o Pacman irá mudar para a direção oposta do fantasma.

Pacman em modo Normal em direção à comida: nesta ação, o Pacman vai determinar as coordenadas da comida mais próxima, e irá na sua direção, usando a mesma estratégia utilizada na ação "Pacman em modo Mega persegue o fantasma" descrita acima.

Conclusão

Em conclusão, apesar de existirem muitas outras estratégias na implementação de um bot para este jogo, foi possível introduzir algumas que se apresentam como essenciais para uma maior fluidez do jogo. Desta forma, os objetivos desta tarefa foram atingidos com sucesso.

Documentation

bot :: Int -> State -> Maybe Play

bot: main bot function

control

Arguments

:: Player
Pacman

-> [Player]
list of all ghosts

-> Maze
maze used to have the coords of foods and walls (when needed)

-> Maybe Play

control: this function controls the Pacman's movements in 3 different situations

aliveGhosts :: [Player] -> [Player]

deadGhosts :: [Player] -> [Player]

chooseByID :: Int -> State -> Player

chooseByID: this function returns the player with the ID and State given

coordsFoodCorridor :: Corridor -> Coords -> [Coords]

coordsFoodCorridor: this function gives the food coordinates of a corridor

coordsFood :: Maze -> [Coords]

coordsFood: this function returns the food coordinates of a maze

coordsGhost

Arguments

:: [Player]
list only with ghosts

-> [Coords]

coordsGhost: this function gives the coordinates of the ghosts

nearestCoords :: Coords -> [Coords] -> Coords

nearestCoords: this function takes a list of coordinates and gives the closest coordinates to the one given

nearestCoords' :: Coords -> [Coords] -> Coords

coordsList :: Coords -> [Coords] -> [Coords]

goToCoords

Arguments

:: Maze
maze used to find if a certain position is a wall or not

-> Player
Player making the move

-> Coords
coords that we want where the Pacman goes

-> Maybe Play

goToCoords: this function returns a play for the player to go in the direction of the given coord

topRightCorner :: Maze -> Player -> Coords -> Maybe Play

topRightCorner: auxiliary function to goToCoords that covers the situations refering to the top right corner of Pacman

topLeftCorner :: Maze -> Player -> Coords -> Maybe Play

topLeftCorner: auxiliary function to goToCoords that covers the situations refering to the top left corner of Pacman

bottomRightCorner :: Maze -> Player -> Coords -> Maybe Play

bottomRightCorner: auxiliary function to goToCoords that covers the situations refering to the bottom right corner of Pacman

bottomLeftCorner :: Maze -> Player -> Coords -> Maybe Play

bottomLeftCorner: auxiliary function to goToCoords that covers the situations refering to the bottom left corner of Pacman

up :: Maze -> Player -> Coords -> Maybe Play

up: auxiliary function to goToCoords that covers the situations when the food or ghosts are above Pacman

down :: Maze -> Player -> Coords -> Maybe Play

down: auxiliary function to goToCoords that covers the situations when the food or ghosts are under Pacman

right :: Maze -> Player -> Coords -> Maybe Play

right: auxiliary function to goToCoords that covers the situations when the food or ghosts are at Pacman's right

left :: Maze -> Player -> Coords -> Maybe Play

left: auxiliary function to goToCoords that covers the situations when the food or ghosts are at Pacman's left

corridor :: Maze -> Player -> Coords -> Maybe Play

corridor: auxiliary function to goToCoords that puts Pacman in the tunnel if he needs it

escapePacman

Arguments

:: Maze
maze used to find if a certain position is a wall or not

-> Player
the pacman if there is a ghost in front of him, in general, he turns to the other side

-> Coords
coords of the nearest ghost

-> Maybe Play

escapePacma: this function gives pacman's plays when he needs to escape from the ghosts

estado :: State

jogadores :: [Player]

jogadorPacman :: Player

fantasma :: Player

labirintoMaze :: Maze

labirintoteste :: State -> Maze

labirintoState :: State

teste :: Maybe Play